+++ packages/flow/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @liteforge/flow
 
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @liteforge/runtime@0.8.0
+
 ## 1.0.0
 
 ### Patch Changes