+++ packages/i18n/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @liteforge/i18n
 
+## 4.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @liteforge/runtime@0.8.0
+
 ## 3.0.0
 
 ### Patch Changes