+++ packages/query/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @liteforge/query
 
+## 5.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @liteforge/runtime@0.8.0
+
 ## 4.0.0
 
 ### Minor Changes