+++ packages/toast/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @liteforge/toast
 
+## 4.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @liteforge/runtime@0.8.0
+
 ## 3.2.0
 
 ### Minor Changes